+++ app/Models/Artist.php
@@ -1,12 +1,19 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
 class Artist extends Model
 {
-    public function tracks() {
+    protected $fillable = [
+        'name',
+    ];
+
+    public function tracks()
+    {
         return $this->belongsToMany(Track::class);
     }
 }