+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Finalizado um ajuste para se bindar o phpcs no projeto. @GustavoSantarosa
 - Criado uma action do github para rodar o pint ao fazer um push. @GustavoSantarosa
 - Iniciado a configuração do swagger. @GustavoSantarosa
+- Criado as models e suas factories. @GustavoSantarosa
 
 # EN
 
@@ -25,3 +26,4 @@
 - Finalized an adjustment to bind phpcs in the project. @GustavoSantarosa
 - Created a GitHub action to run pint on push. @GustavoSantarosa
 - Started configuring Swagger. @GustavoSantarosa
+- Created models and their factories. @GustavoSantarosa